Converted StatefulButton test in typescript, replaced getLocals usage with enzyme utilities.
